Finished Adding interp methods and tests for mu inversion. Mu inversion tests failing
